function

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -9,20 +9,22 @@ int main(void)
 	long int number = 612852475143;
 	int index;
 	int new_idx;
-	int prime;
+	long int prime;
 
-	for (index = 2; index <= number; index++)
+	for (index = 2; index <= number; index++)/*Iteration to check for factors */
 	{
+		/*To check for factors of number*/
 		if (number % index == 0)
 		{
 			for (new_idx = 2; new_idx < index; new_idx++)
 			{
+				/*Checking if factor is not a prime number*/
 				if (index % new_idx == 0)
 				{
 					break;
 				}
 			}
-			if (new_idx == index)
+			if (new_idx == index)/*Checking if the the only divisor of the factor is itself making it a prime*/
 			{
 				prime = index;
 				while (number % index == 0)
@@ -33,6 +35,6 @@ int main(void)
 			}
 		}
 	}
-	printf("%d\n", prime);
+	printf("%ld\n", prime);
 	return (prime);
 }
